[Refactor] Move CliConfig to habitat_common
