+++ components/hab/src/main.rs
@@ -40,8 +40,7 @@ use hab::{cli::{self,
                                      PackageSetFile},
                           list::ListingType,
                           uninstall::UninstallHookMode}},
-          config::{self,
-                   CliConfig},
+          config,
           error::{Error,
                   Result},
           license,
@@ -54,6 +53,7 @@ use hab::{cli::{self,
 use habitat_api_client::BuildOnUpload;
 use habitat_common::{self as common,
                      cli::key_cache_from_matches,
+                     cli_config::CliConfig,
                      command::package::install::{InstallHookMode,
                                                  InstallMode,
                                                  InstallSource,
